add extended resources support to kubermark machine
In kubernetes 1.22.0+, kubemark has support for specifying the resources
for the hollow node using the `--extended-resources` command line flag.
This pull request adds new fields to the KubemarkMachineTemplate CRD
which allows users to customize the resources that their hollow nodes
advertise.

Additionally, this change will create a new default standard for all
hollow nodes of 1 cpu and 4G memory. These can be overridden by adding
custom values to the KubemarkMachineTemplate.

This example demonstrates how to specify the cpu and memory for a hollow
node which will have 32 cpu and 128G memory:
```yaml
apiVersion: infrastructure.cluster.x-k8s.io/v1alpha4
kind: KubemarkMachineTemplate
metadata:
  labels:
    cluster.x-k8s.io/cluster-name: mycluster
  name: kubemark-big-machine
  namespace: default
spec:
  template:
    spec:
      kubemarkOptions:
        extendedResources:
          cpu: "32"
          memory: "128G"
```

The `kubemarkOptions` field is a new struct which is intended to wrap
command line flags for Kubemark. It only contains the
`extendedResources` field currently, but can be expanded with other
options in the future.

The `extendedResources` field is a map which can accept other values to
ensure that users have the maximum flexibility in adding resources.
Although cpu and memory are demonstrated here, a user could add gpu,
networking, or other resources.

Finally, this change updates the README with instructions about the
changes and adds unit tests for the new functions that have been added.