add status.capacity field to KubemarkMachineTemplate
This change adds a status field to KubemarkMachineTemplate, and within
the status, a capacity field to signal the generated machine's sizing.
It is being added to satisfy the optional API requirement for
autoscaling from zero as defined by the enhancement[0]

This change also adds some unit tests for util functions, and refactors
common artifacts into a separate file.

[0]
https://github.com/kubernetes-sigs/cluster-api/blob/main/docs/proposals/20210310-opt-in-autoscaling-from-zero.md