add cert-manager version to test config
this change adds the cert-manager version to the test configuration for
cluster-api. we carry the cert-manager manifest and version along with
our tests so that the local docker registry can be populated with the
images from the upstream cert-manager project instead of having the
images pull when the clusterctl init command runs.