update config secret name to be more clear
it appears that a change in the cluster-api kubeadm bootstrap provider
is causing secrets with the same name as a machine to be created in the
namespace with the machine object. this is causing the kubeconfig secret
created by the capk controller to be overwritten, which cascades into
kubemark not being able to start.

the fix is to add "kubemark-config" as a suffix to secret.